Run dev environment with Gunicorn and NGINX
This will be used only as a reference for the future and will be
undone, as for development it's faster to use the Django
development server instead of setting up a complex stack.

Collects static data into a shared volume with /static location
between Django and NGINX starts a Gunicorn and NGINX server.

NGINX acts both as a static file server and reverse proxy.
That means that NGINX will decide depending on the request if to
get the data from the Gunicorn server or serve the content itself
(static files). The reason of setting up NGINX to serve static
data is because Gunicorn is not suitable for that.