Eliminate downloading most components
From a _desktop_ setup point of view, it's fine to download a bunch
of files manually because you only do it once. Still have to configure
a client and start the server and run the suite for each browser of
interest. This change adds a WDIO service that starts and stop the
server automatically and also run the test suite for each configured
browser. This means that users/devs do not need to download the server
and drivers separately and do not need to ensure the server is running
before the tests starts. The service does this for you.