Fix urllib3 vulnerability
CVE-2018-20060 More information
high severity
Vulnerable versions: < 1.23
Patched version: 1.23
urllib3 before version 1.23 does not remove the Authorization HTTP header when following a cross-origin redirect (i.e., a redirect that differs in host, port, or scheme). This can allow for credentials in the Authorization header to be exposed to unintended hosts or transmitted in cleartext.